fix: address ADK recipe review feedback
