+++ contrib/python/amg-memory-guard/amg_memory_guard_adk/agent.py
@@ -1,3 +1,15 @@
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """A minimal ADK agent that guards a durable session-state boundary."""
 
 from __future__ import annotations
@@ -25,13 +37,15 @@ def remember_external_observation(
     if tool_context is None:
         return {"status": "error", "message": "ADK tool context is required."}
 
+    session_id = tool_context._invocation_context.session.id
+
     try:
         action = guard.write(
             "adk.external_observation",
             content,
             source="adk.external_tool",
             source_class=SourceClass.EXTERNAL_TOOL,
-            task_id="adk-session",
+            task_id=session_id,
         )
     except PolicyViolation as exc:
         return {
@@ -49,7 +63,7 @@ def remember_external_observation(
 
 root_agent = Agent(
     name="guarded_memory_agent",
-    model=os.getenv("MODEL_NAME", "gemini-3.5-flash"),
+    model=os.getenv("MODEL_NAME"),
     instruction=(
         "You are a concise research assistant. Use remember_external_observation "
         "only for compact, factual external observations that will help a later turn."
