+++ contrib/python/amg-memory-guard/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "amg-memory-guard-adk"
+name = "amg-memory-guard"
 version = "0.1.0"
 description = "Guard Google ADK durable session-state writes with OWASP Agent Memory Guard."
 readme = "README.md"
@@ -25,3 +25,7 @@ packages = ["amg_memory_guard_adk"]
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
+
+[[tool.uv.index]]
+url = "https://pypi.org/simple/"
+default = true
